requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pyserial==3.5
qt-material==2.14
requests==2.31.0
PyGithub==2.1.1
semantic-version==2.10.0 